pyproject.toml
[project]
name = "final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-dartmouth>=0.3.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
]
